backend/requirements.txt
fastapi
mcrcon
passlib
psutil
pydantic
python_jose
Requests
rich
uvicorn
